feat: add some more tests for tag usages
